Setting charging status
Learned from this page:
http://developer.android.com/training/monitoring-device-state/battery-monitoring.html

+++ src/app/src/main/java/com/appsoluteperfection/chargedreboot/domain/DeviceStatus.java
@@ -2,15 +2,15 @@ package com.appsoluteperfection.chargedreboot.domain;
 
 public class DeviceStatus {
 
-    private int _batteryLevel;
+    private float _batteryLevel;
     private boolean _isCharging;
 
-    public DeviceStatus(int batteryLevel, boolean isCharging) {
+    public DeviceStatus(float batteryLevel, boolean isCharging) {
         _batteryLevel = batteryLevel;
         _isCharging = isCharging;
     }
 
-    public int getBatteryLevel(){
+    public float getBatteryLevel(){
         return _batteryLevel;
     }
